+++ internal/infrastructure/repository/model_test.go
@@ -3,19 +3,16 @@ package repository
 import (
 	"testing"
 
-	"personal-finance/internal/domain"
+	"personal-finance/internal/domain/fixture"
 
 	"github.com/stretchr/testify/assert"
 )
 
 func TestToMovementModel(t *testing.T) {
-	// Arrange
-	domainMovement := domain.MovementMock()
+	domainMovement := fixture.MovementMock()
 
-	// Act
 	dbModel := ToMovementModel(domainMovement)
 
-	// Assert
 	assert.Equal(t, *domainMovement.ID, *dbModel.ID)
 	assert.Equal(t, domainMovement.Description, dbModel.Description)
 	assert.Equal(t, domainMovement.Amount, dbModel.Amount)
@@ -28,14 +25,11 @@ func TestToMovementModel(t *testing.T) {
 }
 
 func TestMovementModelToDomain(t *testing.T) {
-	// Arrange
-	domainMovement := domain.MovementMock()
+	domainMovement := fixture.MovementMock()
 	dbModel := ToMovementModel(domainMovement)
 
-	// Act
 	resultDomain := dbModel.ToDomain()
 
-	// Assert
 	assert.Equal(t, *domainMovement.ID, *resultDomain.ID)
 	assert.Equal(t, domainMovement.Description, resultDomain.Description)
 	assert.Equal(t, domainMovement.Amount, resultDomain.Amount)
@@ -45,4 +39,31 @@ func TestMovementModelToDomain(t *testing.T) {
 	assert.Equal(t, *domainMovement.WalletID, *resultDomain.WalletID)
 	assert.Equal(t, *domainMovement.CategoryID, *resultDomain.CategoryID)
 	assert.Equal(t, domainMovement.TypePaymentID, resultDomain.TypePaymentID)
+}
+
+func TestToSubCategoryModel(t *testing.T) {
+	domainSubCategory := fixture.SubCategoryMock()
+
+	dbModel := ToSubCategoryModel(domainSubCategory)
+
+	assert.Equal(t, *domainSubCategory.ID, *dbModel.ID)
+	assert.Equal(t, domainSubCategory.Description, dbModel.Description)
+	assert.Equal(t, domainSubCategory.UserID, dbModel.UserID)
+	assert.Equal(t, *domainSubCategory.CategoryID, *dbModel.CategoryID)
+	assert.Equal(t, domainSubCategory.DateCreate, dbModel.DateCreate)
+	assert.Equal(t, domainSubCategory.DateUpdate, dbModel.DateUpdate)
+}
+
+func TestSubCategoryModelToDomain(t *testing.T) {
+	domainSubCategory := fixture.SubCategoryMock()
+	dbModel := ToSubCategoryModel(domainSubCategory)
+
+	resultDomain := dbModel.ToDomain()
+
+	assert.Equal(t, *domainSubCategory.ID, *resultDomain.ID)
+	assert.Equal(t, domainSubCategory.Description, resultDomain.Description)
+	assert.Equal(t, domainSubCategory.UserID, resultDomain.UserID)
+	assert.Equal(t, *domainSubCategory.CategoryID, *resultDomain.CategoryID)
+	assert.Equal(t, domainSubCategory.DateCreate, resultDomain.DateCreate)
+	assert.Equal(t, domainSubCategory.DateUpdate, resultDomain.DateUpdate)
 }